pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-evaluate"
version = "0.1.0"
description = "Engineering-focused evaluation framework for open-source LLM/VLM models and datasets"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Llm_Evaluate Team" }]
dependencies = [
  "typer>=0.12.0",
  "pyyaml>=6.0.0",
  "datasets>=2.19.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
modelscope = ["modelscope>=1.14.0"]
transformers = ["transformers>=4.45.0", "torch>=2.2.0", "pillow>=10.0.0"]
all = [
  "modelscope>=1.14.0",
  "transformers>=4.45.0",
  "torch>=2.2.0",
  "pillow>=10.0.0",
]
dev = ["pytest>=8.2.0", "ruff>=0.5.0"]

[project.scripts]
llm-evaluate = "llm_evaluate.cli.app:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
